backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-jose[cryptography]>=3.3.4
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.1
google-cloud-aiplatform>=1.42.1
google-cloud-storage>=2.14.0
google-cloud-firestore>=2.14.0
pyopenssl
pillow>=10.0.0
google-cloud-texttospeech>=2.16.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
